Feature: Add Header component to appear on detail story page
