+++ src/Card/Card.js
@@ -3,14 +3,18 @@ import { Link } from 'react-router-dom';
 
 function Card({ article }) {
   return (
-    <Link to={`/${article.id}/${article.title}`}>
+      
       <div className="card">
-        <img src={article.urlToImage}alt="article photo"></img>
-        <h3>{article.title}</h3>
-        <p>{article.formattedDate} —{article.description}</p>
-        <p className="type">| {article.type}</p>
-      </div>
-      </Link>
+         <p className="type">| {article.type}</p>
+        <Link to={`/${article.id}/${article.title}`}>
+          <div className="image-wrapper">
+            <img src={article.urlToImage} alt="article photo" />
+            <div className="overlay"></div>
+          </div>
+          <h3>{article.title}</h3>
+          <p>{article.formattedDate} —{article.description}</p>
+        </Link>
+      </div>   
   );
 }
