+++ src/Header/Header.css
@@ -1,0 +1,26 @@
+.header {
+  background-color: #FFFFFF;
+  min-height: 10vh;
+  display: flex;
+  flex-direction: row;
+  align-items: flex-end;
+  justify-content: space-between;
+  padding: 3%;
+  font-size: calc(2px + 1.5vmin);
+  color: rgb(2, 34, 2);
+}
+
+.App-logo {
+  height: 25vmin;
+  pointer-events: all;
+  padding-left: 5%;
+}
+
+.date-container {
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.today {
+  color: #010101;
+}
