+++ src/StoryDetail/StoryDetail.js
@@ -1,5 +1,7 @@
 import "./StoryDetail.css";
 import { useParams } from "react-router-dom";
+import { Link } from 'react-router-dom';
+
 
 function StoryDetail({ worldNews }) {
   let { id } = useParams();
@@ -14,13 +16,18 @@ function StoryDetail({ worldNews }) {
   }
   return (
     <main>
-    <div className="story-detail">
-      <h4>| {article.type}</h4>
-      <img src={article.urlToImage} alt="Image illustrating the article"></img>
-      <h2>{article.title}</h2>
-      <h5>{article.source.name}</h5>
-      <p>{article.formattedDate} —{article.content}</p>
-    </div>
+      <div className="back-link">
+        <Link to={"/"}>
+        <h4 >⬅︎ BACK TO HOME</h4>
+        </Link>
+      </div>
+      <div className="story-detail">
+        <h4>| {article.type}</h4>
+        <img src={article.urlToImage} alt="Image illustrating the article"></img>
+        <h2>{article.title}</h2>
+        <h5>{article.source.name}</h5>
+        <p>{article.formattedDate} —{article.content}</p>
+      </div>
     </main>
   );
 }
